fix: int64 -> int32

+++ modules/api/api.go
@@ -199,7 +199,7 @@ func (c *Caller) call(action string, spec *onebot.Spec, p Getter) global.MSG {
 	case "get_group_file_url":
 		p0 := p.Get("group_id").Int()
 		p1 := p.Get("file_id").String()
-		p2 := p.Get("[busid,bus_id].0").Int()
+		p2 := int32(p.Get("[busid,bus_id].0").Int())
 		return c.bot.CQGetGroupFileURL(p0, p1, p2)
 	case "upload_group_file":
 		p0 := p.Get("group_id").Int()
@@ -224,7 +224,7 @@ func (c *Caller) call(action string, spec *onebot.Spec, p Getter) global.MSG {
 	case "delete_group_file":
 		p0 := p.Get("group_id").Int()
 		p1 := p.Get("file_id").String()
-		p2 := p.Get("[busid,bus_id].0").Int()
+		p2 := int32(p.Get("[busid,bus_id].0").Int())
 		return c.bot.CQGroupFileDeleteFile(p0, p1, p2)
 	}
 	return coolq.Failed(404, "API_NOT_FOUND", "API不存在")
